+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Fulltext query on `str` fields in the database when using `q` filter.
 
 ## [1.1.2] - 2024-06-10